IndexedDB/tests/submissions/Opera/Microsoft-port/idbobjectstore_add14.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBObjectStore.add() - Attempt to add a record where a value being indexed does not meet the constraints of a valid key </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support.js"></script>

<script>
    var db,
      t = async_test(),
      record = { key: 1, indexedProperty: { property: "data" } };

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;

        var rq,
          objStore = db.createObjectStore("store", { keyPath: "key" });

        objStore.createIndex("index", "indexedProperty");

        assert_throws("DataError",
            function() { rq = objStore.add(record); });

        assert_equals(rq, undefined);
        t.done();
    };
</script>

<div id="log"></div>
